* 'master' of https://github.com/actiontech/dble: (440 commits)
  develop notice
  develop notice
  develop notice
  reload cluster error
  typo
  fix first start warnning log
  derived subquery is global table #685
  load data lost conn #683
  table name #683
  Table not exist #683
  load data no local empty file #683
  typo
  charset default value #689
  typo
  treat single node sharding table as no-sharding table #687
  join with null #681 #634
  optimizer er-global subquery
  load data with empty file get error return (#684) #683
  function charset #679
  Union fix (#672)
  ...

+++ src/main/java/com/actiontech/dble/statistic/stat/SqlResultSizeRecorder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016-2017 ActionTech.
+ * Copyright (C) 2016-2018 ActionTech.
  * License: http://www.gnu.org/licenses/gpl.html GPL version 2 or higher.
  */
 
@@ -18,7 +18,7 @@ public class SqlResultSizeRecorder {
     private ConcurrentMap<String, SqlResultSet> sqlResultSetMap = new ConcurrentHashMap<>();
 
 
-    public void addSql(String sql, int resultSetSize) {
+    public void addSql(String sql, long resultSetSize) {
         SqlResultSet sqlResultSet;
         SqlParser sqlParserHigh = new SqlParser();
         sql = sqlParserHigh.mergeSql(sql);
